@roryp2 Refactor runners to merge browserstack and local one to remove code duplication
