+++ lib/submit.js
@@ -34,6 +34,7 @@ module.exports = function (scholarUrl, imageResults, callback) {
                         resp.name = image.scenario.name;
                         return submissionCallback(err, resp);
                     } catch (e) {
+                        console.error(`Error! ${e} occurred. Scholar response was: ${body}`);
                         submissionCallback(e);
                     }
                 });
